Fatras: combine particle selectors into single header
