+++ Tests/UnitTests/Fatras/Selectors/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(unittest_extra_libraries ActsFatras)
 
-add_unittest(FatrasChargeSelectors ChargeSelectorsTests.cpp)
 add_unittest(FatrasKinematicCasts KinematicCastsTests.cpp)
-add_unittest(FatrasPdgSelectors PdgSelectorsTests.cpp)
+add_unittest(FatrasParticleSelectors ParticleSelectorsTests.cpp)
 add_unittest(FatrasSelectorHelpers SelectorHelpersTests.cpp)
